extract StitchBlobs MPI exchange and debug utilities into headers
